backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4

# Database & Auth
supabase==2.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# HTTP & Web
httpx==0.28.1
python-multipart==0.0.18
beautifulsoup4==4.12.3

# AI & ML
google-generativeai==0.8.3
ollama==0.4.4

# Voice Service
websockets==14.1
aiofiles==24.1.0
numpy>=1.26.0

# Optional: Local voice models (uncomment if needed)
# faster-whisper==1.0.0
# TTS==0.22.0
# torch==2.0.1

# Serverless
mangum==0.19.0
